empleados varios planteles, denominacion, planes de estudio

+++ database/migrations/2020_03_19_122518_add_coidigo_creditos_to_materia_table.php
@@ -15,7 +15,7 @@ class AddCoidigoCreditosToMateriaTable extends Migration
     {
         Schema::table('materia', function (Blueprint $table) {
             $table->string('codigo')->nullable();
-            $table->integer('creditos')->nullable();
+            $table->decimal('creditos')->nullable();
         });
     }
